Looks like it correctly re-implements the python handwring example

+++ type.h
@@ -3,7 +3,7 @@
 
 #include <vector>
 
-typedef float fpt;
+typedef double fpt;
 typedef std::vector<fpt> fpt_vect;
 
 #endif // TYPE_H
